Not a fully working Docker setup just yet. Just some setup for eventual ghrc.io and package releases for actual docker testing, fixing, and eventual personal migration of the bot to an unraid server. On top of that it has some documentation updates for manga.py along with a new command to explain a particular bugfix that will probably affect actual users. On that note this also contains some bug fixes both in manga.py and in MangaDexImplementation.py, some of these fixes more temporary than others as it's just a band-aid for still undeveloped features. And lastly some README.md updates but that's small.

+++ bot/exts/manga.py
@@ -1,3 +1,4 @@
+import asyncio
 import os
 
 import discord
@@ -17,38 +18,49 @@ class mangaCommands(commands.Cog):
         self.bot = bot
         self.md = mdAPI()
 
-    # @commands.command(name="SearchMDex")
-    @app_commands.command(name="searchmdex")
+    @app_commands.command(name="searchmdex", description="Search for Manga from MangaDex right from your discord")
     @app_commands.describe(title="Title of the mangga")
     async def SearchMDex(self, interaction: discord.Interaction, title: str) -> None:
         """
-        Search manga dex for a user specified manga and return to them an embedd containing a link and other info
+        Search MangaDex for a user specified manga and return to them an embedd containing a link and other info
         about the manga
 
         :param interaction: Necessary for slash commands to function properly
         :param title: Title of the manga to be searched for, passed in by user
         """
-        # md = mdAPI()
         mangaJson = self.md.getMangaJson(title)
 
-        path = self.md.getMangaCover(mangaJson=mangaJson)
-        embed = self.md.createDiscordEmbed(mangaJson=mangaJson)
+        if mangaJson is None:
+            await interaction.response.send_message("Something went wrong. Here are some possible reasons:\n"  # noqa
+                                                    "1. The manga youre looking for does not exist on MDex's site\n"
+                                                    "2. Something about the spelling is wrong\n"
+                                                    "3. Its MDex's fault\n"
+                                                    "4. You searched for Pokemon\n")
+        else:
+            path = self.md.getMangaCover(mangaJson=mangaJson)
+            embed = self.md.createDiscordEmbed(mangaJson=mangaJson)
 
-        # There's a strange bug on library end where the image isnt being set with the actual url retrieved from the
-        # MangaDex API for now iamages are created and sent alongside the embed before having their files closed and
-        # deleted by the OS
-        file = discord.File(path, filename="cover.jpg")
-        embed.set_image(url="attachment://cover.jpg")
+            # There's a strange bug on library end where the image isnt being set with the actual url retrieved from the
+            # MangaDex API for now iamages are created and sent alongside the embed before having their files closed and
+            # deleted by the OS
+            file = discord.File(path, filename="cover.jpg")
+            embed.set_image(url="attachment://cover.jpg")
 
-        # Pycharm marks this as an unresolved reference but it both shows up in the API documentation and works so
-        await interaction.response.send_message(embed=embed, file=file)  # noqa
-        file.close()
-        os.remove(path)
+            # Pycharm marks this as an unresolved reference but it both shows up in the Library documentation and
+            # works so
+            await interaction.response.send_message(embed=embed, file=file)  # noqa
+            file.close()
+            os.remove(path)
 
     # At some point itll probably be a good idea to make a helper method to handle most of whats going on here
     # I'll come back to it later
-    @app_commands.command(name="randomanga")
-    async def RandomManga(self, interaction: discord.Interaction):
+    @app_commands.command(name="randomanga", description="Get a Random Manga from MangaDex, never know whats out there")
+    async def RandomManga(self, interaction: discord.Interaction) -> None:
+        """
+        Search for a random manga from MangaDex and return it to the user
+
+        @param interaction:  Necessary for slash commands to function properly
+        """
         mangaJson = self.md.getRandomManga()
 
         path = self.md.getMangaCover(mangaJson=mangaJson)
@@ -58,6 +70,27 @@ class mangaCommands(commands.Cog):
         file = discord.File(path, filename="cover.jpg")
         embed.set_image(url="attachment://cover.jpg")
 
-        await interaction.response.send_message(embed=embed, file=file)  # noqa
+        # Preemptive measures to avoid unkown interaction errors
+        await interaction.response.defer()  # noqa
+        await asyncio.sleep(3)
+        await interaction.followup.send(embed=embed, file=file)
+
         file.close()
-        os.remove(path)
+
+    # if this is still here in 2024 someone bother me to fix it
+    @app_commands.command(name="pokemon", description="If you're wondering why you cant search Pokemon Manga")
+    async def WhyWereNotAllowedToSearchForPokemonRN(self, interaction: discord.Interaction) -> None:
+        """
+        Just an explanation as to why Pokemon searches are currently being filtered out, idk why Im writing this
+        docstring tbh
+
+        @param interaction:  Necessary for slash commands to function properly
+        """
+        await interaction.response.send_message("If you're here you're probably wondering \"Why cant we lookup "  # noqa
+                                                "Pokemon?\"\nPokemon searches on MDex's API seem to be somewhat "
+                                                "broken. Using the API to search up Pokemon will bring up various "
+                                                "Pokemon spinoffs, not the original series. Meanhwile searching on "
+                                                "MDex's site will bring up the original Pokemon but none of those "
+                                                "spinoffs. Because of this all Pokemon searches are disabled until a "
+                                                "proper fix can be implemented. If this still isnt fixxed in 2024 find"
+                                                " someone who knows and will bother my creator about it.")
